requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pypdf==4.0.1
pandas==2.2.3
numpy==1.26.4
streamlit==1.30.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6




# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4
# pypdf==3.17.1
# pandas==2.1.3
# numpy==1.25.2
# streamlit==1.28.1
# python-dotenv==1.0.0
# pydantic==2.5.0
# python-multipart==0.0.6